Fix bugs related to deleting CardDataImage rows.
Extension apps and/or Catapult could crash because IDs were not handled
correctly. Fixed that. Removed extra '/' in Uri path constants. Fixed id
parsing in PublishableCard.

+++ sdk/src/org/cyanogenmod/launcher/home/api/cards/PublishableCard.java
@@ -90,7 +90,7 @@ public abstract class PublishableCard {
 
             if (result != null) {
                 // Store the resulting ID
-                setId(Integer.parseInt(result.getLastPathSegment()));
+                setId(Long.parseLong(result.getLastPathSegment()));
             }
         }
     }
